Improved event logs for JSTPublisher ONLY. JBIDE-1047
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@4272 a97e2381-89e5-4abb-bab3-167db6db766c

+++ as/plugins/org.jboss.ide.eclipse.as.core/jbosscore/org/jboss/ide/eclipse/as/core/util/FileUtil.java
@@ -48,20 +48,25 @@ public class FileUtil {
 					safeDelete(children[i], listener);
 				}
 			}
-			try {
-				boolean tmp = file.delete();
-				if( listener != null ) listener.folderDeleted(file, tmp, null);
-			} catch( SecurityException sex) {
-				if( listener != null ) listener.folderDeleted(file, false, sex);
+			
+			if( file.exists()) {
+				try {
+					boolean tmp = file.delete();
+					if( listener != null ) listener.folderDeleted(file, tmp, null);
+				} catch( SecurityException sex) {
+					if( listener != null ) listener.folderDeleted(file, false, sex);
+				}
 			}
 		}
 		
 		// files only
-		try {
-			boolean tmp = file.delete();
-			if( listener != null ) listener.fileDeleted(file, tmp, null);
-		} catch( SecurityException sex) {
-			if( listener != null ) listener.fileDeleted(file, false, sex);
+		if( file.exists() ) {
+			try {
+				boolean tmp = file.delete();
+				if( listener != null ) listener.fileDeleted(file, tmp, null);
+			} catch( SecurityException sex) {
+				if( listener != null ) listener.fileDeleted(file, false, sex);
+			}
 		}
 	}
